+++ mypy/test/testdaemon.py
@@ -44,7 +44,9 @@ def test_daemon(testcase: DataDrivenTestCase) -> None:
         cmd = cmd[1:].strip()
         cmd = cmd.replace("{python}", sys.executable)
         if cmd.split()[1] in ("start", "restart", "run"):
-            cmd = cmd.replace("-- ", "-- --legacy ")
+            cmd = cmd.replace(
+                "-- ", "-- --no-strict --no-infer-function-types --no-default-return "
+            )
         sts, output = run_cmd(cmd)
         output_lines = output.splitlines()
         output_lines = normalize_error_messages(output_lines)