+++ mypy/test/testpep561.py
@@ -108,7 +108,7 @@ def test_pep561(testcase: DataDrivenTestCase) -> None:
                     f.write(f"{s}\n")
             cmd_line.append(program)
 
-        cmd_line.extend(["--no-error-summary", "--hide-error-codes", "--legacy"])
+        cmd_line.extend(["--no-error-summary", "--hide-error-codes", "--no-strict"])
         if python_executable != sys.executable:
             cmd_line.append(f"--python-executable={python_executable}")