+++ mypyc/test/test_commandline.py
@@ -52,7 +52,7 @@ class TestCommandLine(MypycDataSuite):
         try:
             # Compile program
             cmd = subprocess.run(
-                [sys.executable, "-m", "mypyc", "--legacy", *args],
+                [sys.executable, "-m", "mypyc", "--no-strict", *args],
                 stdout=subprocess.PIPE,
                 stderr=subprocess.STDOUT,
                 cwd="tmp",