+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Added
 - Deprecate python 3.7 support (#457)
 - Allow denotation of tuple types with tuple literals (#458)
+- Removed `--legacy` flag in favour of `--no-strict` (#445)
+- `default-return` is now enabled by default (#445)
 ### Enhancements
 - Removed `Any` from the typings for `re.Match` group functions. (#459)
 ### Fixes